Move to `src`-based project
