+++ src/roguelike_game/Assets/Clear Block.png
