+++ src/roguelike_game/Assets/Grey Block.jpg
