+++ src/roguelike_game/Assets/Transparent Border.png
