+++ src/roguelike_game/__init__.py
@@ -1,0 +1,4 @@
+from .game import main
+
+if __name__ == "__main__":
+    main()
